Fixed all known bugs.
Code-block #5, Line #52 makes strategy percentages to accumulate, by carrying forward the strategy percentage values across (even new) graphs. Not clear on whether this part of the experiment is correct. 
If this line is commented, the strategy percentages fall back to the default value in code-block #1 for each new graph. If the experiment is done this way, the strategy percentages just fluctuates around their default values.

Direction to work on: The experiment currently plots how the strategy percentage values change cumulatively (even across graphs!) and this doesn't feel right. Shouldn't it be rather checking how the values change across iterations in one graph? Where is this in conflict with the Monte-Carlo design?